Authentication & Authorization on post delete api

+++ controllers/api/v1/posts_api.js
@@ -45,12 +45,23 @@ module.exports.getById = function (request, response) {
         });
 };
 
+// /=/ Update post by ID Request Controller for (/api/v1/post/{id}) [promise-then]
+// module.exports.update = function (request, response) {
+//     if (request.user.id == request.params.id) {
+//         Post.findByIdAndUpdate(request.body.id)
+//         .then(post => { })
+//         .catch(error => { });
+//     } else {
+//         /=/ handle error here.
+//     }
+// }
+
 // DELETE Request Controller for (/api/v1/posts/{id}) [async-await]
-module.exports.destroy = async function (request, response) {
+module.exports.delete = async function (request, response) {
     try { 
         let post = await Post.findById(request.params.id);
 
-        // if (post.user == request.user.id) {
+        if (post.user == request.user.id) {
             post.deleteOne(); // Replace post.remove() with post.deleteOne()
 
             await Comment.deleteMany({ post: request.params.id });
@@ -59,11 +70,11 @@ module.exports.destroy = async function (request, response) {
                 message: "Post and associated comments Deleted Successfully.",
                 post: post,
             });
-        // } else {
-        //     console.log('Unable to delete the post ::', post);
-        //     response.flash('error', 'Unable to delete the Post.');
-        //     return response.redirect('back');
-        // }
+        } else {
+            return response.status(401).json({ 
+                message: "You are Unauthorized! You cannot delete this post.",
+            });
+        }
     } catch (error) { 
         console.error('Delete Post Error on API call :', error);
         return response.status(500).json({
